Search working, results displaying as required, links to profiles working. Still need to add avatar
